+++ tests/test_player.py
@@ -1,4 +1,5 @@
 import os
+import vlc
 import pytest
 import castero.player as player
 
@@ -13,12 +14,16 @@ def test_player_init():
 def test_player_play():
     myplayer = player.Player("MLK Dream", my_dir + "/media/MLK_Dream_10s.mp3")
     myplayer.play()
+    while myplayer._player.get_state() == vlc.State.Opening:
+        pass
     assert myplayer.state == 1
 
 
 def test_player_pause():
     myplayer = player.Player("MLK Dream", my_dir + "/media/MLK_Dream_10s.mp3")
     myplayer.play()
+    while myplayer._player.get_state() != vlc.State.Playing:
+        pass
     myplayer.pause()
     assert myplayer.state == 2
 
@@ -26,6 +31,8 @@ def test_player_pause():
 def test_player_stop():
     myplayer = player.Player("MLK Dream", my_dir + "/media/MLK_Dream_10s.mp3")
     myplayer.play()
+    while myplayer._player.get_state() != vlc.State.Playing:
+        pass
     myplayer.stop()
     assert myplayer.state == 0
 
@@ -40,6 +47,8 @@ def test_player_del():
 def test_player_seek():
     myplayer = player.Player("MLK Dream", my_dir + "/media/MLK_Dream_10s.mp3")
     myplayer.play()
+    while myplayer._player.get_state() != vlc.State.Playing:
+        pass
     myplayer.seek(1, 10)
     assert myplayer.time == 10 * 1000
 
@@ -52,10 +61,20 @@ def test_player_title():
 def test_player_time():
     myplayer = player.Player("MLK Dream", my_dir + "/media/MLK_Dream_10s.mp3")
     myplayer.play()
+    while myplayer._player.get_state() != vlc.State.Playing:
+        pass
     assert myplayer.time == 0
 
 
 def test_player_time_str():
     myplayer = player.Player("MLK Dream", my_dir + "/media/MLK_Dream_10s.mp3")
     myplayer.play()
+    while myplayer._player.get_state() != vlc.State.Playing:
+        pass
     assert myplayer.time_str == "00:00:00/00:00:05"
+
+
+def test_player_create_error():
+    with pytest.raises(player.PlayerCreateError):
+        myplayer = player.Player("fake", "not a real path")
+        myplayer._create_player()